refactor: home codes
* refactor home code structure
* refactor methods
